+++ controllers/admin/IndexController.php
@@ -406,6 +406,19 @@ class IndexController extends AdminBase
                     case Stylesheet::TYPE_FONT:
                         $stylesheet->$key = $settings[$key];
                         break;
+                    case Stylesheet::TYPE_IMAGE:
+                        if (isset($settings[$key]) && $settings[$key]) {
+                            $stylesheet->$key = $settings[$key];
+                        } elseif (isset($_FILES['uploaded_' . $key]) && $_FILES['uploaded_' . $key]) {
+                            try {
+                                $file = ConsultationFile::uploadImage($this->consultation, 'uploaded_' . $key);
+
+                                $stylesheet->$key = $file->getUrl();
+                            } catch (FormError $e) {
+                                \yii::$app->session->setFlash('error', $e->getMessage());
+                            }
+                        }
+                        break;
                 }
             }
             $siteSettings->setStylesheet($stylesheet);