+++ views/admin/index/theming.php
@@ -30,7 +30,11 @@ foreach (Stylesheet::getAllSettings() as $key => $setting) {
 ?>
 <h1><?= \Yii::t('admin', 'theme_title') ?></h1>
 
-<?= Html::beginForm(UrlHelper::createUrl('/admin/index/theming'), 'POST', ['class' => 'themingForm']) ?>
+<?= Html::beginForm(UrlHelper::createUrl('/admin/index/theming'), 'POST', [
+    'class'                    => 'themingForm',
+    'enctype'                  => 'multipart/form-data',
+    'data-antragsgruen-widget' => 'backend/Theming'
+]) ?>
 
 <?php
 foreach ($settingsByBlock as $group => $settings) {
@@ -41,7 +45,7 @@ foreach ($settingsByBlock as $group => $settings) {
         foreach ($settings as $key => $setting) {
             $title = \Yii::t('admin', 'theme_' . $key);
             ?>
-            <tr>
+            <tr class="row_<?= $setting['type'] ?>">
                 <th>
                     <label for="stylesheet-<?= Html::encode($key) ?>"><?= Html::encode($title) ?></label>
                 </th>
@@ -79,6 +83,58 @@ foreach ($settingsByBlock as $group => $settings) {
                         echo Html::checkbox('stylesheet[' . $key . ']', $stylesheet->getValue($key), [
                             'id' => 'stylesheet-' . $key
                         ]);
+                    }
+                    if ($setting['type'] === Stylesheet::TYPE_IMAGE) {
+                        ?>
+                        <div class="logoPreview">
+                            <?php
+                            if ($stylesheet->getValue($key)) {
+                                echo '<img alt="" src="' . Html::encode($stylesheet->getValue($key)) . '">';
+                            }
+                            ?>
+                        </div>
+                        <div class="imageChooser">
+                            <input type="hidden" name="stylesheet[<?= $key ?>]"
+                                   value="<?= Html::encode($stylesheet->getValue($key)) ?>" autocomplete="off">
+                            <div class="uploadCol">
+                                <input type="file" name="uploaded_<?= $key ?>" class="form-control"
+                                       id="upload_<?= $key ?>">
+                                <label for="upload_<?= $key ?>">
+                                    <span class="glyphicon glyphicon-upload"></span>
+                                    <span class="text"
+                                          data-title="<?= Html::encode(\Yii::t('admin', 'theme_img_upload')) ?>">
+                                        <?= \Yii::t('admin', 'theme_img_upload') ?>
+                                    </span>
+                                </label>
+                            </div>
+                            <?php
+                            $images = $consultation->site->getFileImages();
+                            if (count($images) > 0) {
+                                ?>
+                                <div class="dropdown imageChooserDd">
+                                    <button class="btn btn-default dropdown-toggle" type="button"
+                                            id="fileChooseDropdownBtn_<?= $key ?>"
+                                            data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
+                                        <?= \Yii::t('admin', 'con_logo_url_choose') ?>
+                                        <span class="caret"></span>
+                                    </button>
+                                    <div class="dropdown-menu dropdown-menu-right"
+                                         aria-labelledby="fileChooseDropdownBtn_<?= $key ?>">
+                                        <ul>
+                                            <?php
+                                            foreach ($images as $file) {
+                                                $src = Html::encode($file->getUrl());
+                                                echo '<li><a href="#"><img alt="" src="' . $src . '"></a></li>';
+                                            }
+                                            ?>
+                                        </ul>
+                                    </div>
+                                </div>
+                                <?php
+                            }
+                            ?>
+                        </div>
+                        <?php
                     }
                     ?>
                 </td>