+++ History.md
@@ -2,7 +2,7 @@
 
 ## Version 4.2.0 [not released yet]
 
-- It is now possible to modify the layout of Antragsgrün, e.g. by choosing custom colors. To create a custom theme, go to "Settings" -> "This consultation" -> "Custom theme".
+- It is now possible to modify the layout of Antragsgrün, e.g. by choosing custom colors or uploading a background image. To create a custom theme, go to "Settings" -> "This consultation" -> "Custom theme".
 - When logged in, users can now add private notes to each motion, amendment and motion paragraph.
 - Improvements regarding uploaded images:
   - Images uploaded at one consultation can now be used in other consultations of the same site as well.